Renamed to structured-event-log
- Renamed interface to StructuredEventLog
- Moved subinterfaces out of class
